ehr/requirements.txt
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
absl-py==1.4.0
asttokens==2.2.1
chex==0.1.7
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dm-haiku==0.0.9
dm-tree==0.1.8
executing==1.2.0
femr-cuda==0.0.198
fonttools==4.40.0
icecream==2.1.3
jax==0.4.8
jaxlib==0.4.7+cuda11.cudnn86
jmp==0.0.4
joblib==1.2.0
kiwisolver==1.4.4
lightgbm==3.3.5
loguru==0.7.0
matplotlib==3.7.1
ml-dtypes==0.2.0
msgpack==1.0.5
nptyping==2.4.1
numpy==1.24.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
opt-einsum==3.3.0
optax==0.1.4
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
threadpoolctl==3.1.0
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
zstandard==0.21.0
